+++ platform/execution-impl/src/com/intellij/execution/impl/RunManagerImpl.kt
@@ -1,10 +1,12 @@
-// Copyright 2000-2021 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+// Copyright 2000-2022 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
+@file:Suppress("ReplaceGetOrSet", "ReplacePutWithAssignment")
+
 package com.intellij.execution.impl
 
-import com.intellij.ProjectTopics
 import com.intellij.configurationStore.*
 import com.intellij.execution.*
 import com.intellij.execution.configurations.*
+import com.intellij.execution.runToolbar.RunToolbarSlotManager
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.execution.runners.ExecutionUtil
 import com.intellij.execution.runners.ProgramRunner
@@ -14,6 +16,7 @@ import com.intellij.ide.util.PropertiesComponent
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ModalityState
+import com.intellij.openapi.application.ReadAction
 import com.intellij.openapi.components.*
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.diagnostic.runAndLogException
@@ -24,8 +27,6 @@ import com.intellij.openapi.options.SchemeManagerFactory
 import com.intellij.openapi.project.IndexNotReadyException
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.impl.ProjectManagerImpl
-import com.intellij.openapi.roots.ModuleRootEvent
-import com.intellij.openapi.roots.ModuleRootListener
 import com.intellij.openapi.startup.StartupManager
 import com.intellij.openapi.updateSettings.impl.pluginsAdvertisement.UnknownFeaturesCollector
 import com.intellij.openapi.util.ClearableLazyValue
@@ -36,17 +37,26 @@ import com.intellij.openapi.util.registry.Registry
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.project.isDirectoryBased
 import com.intellij.util.*
-import com.intellij.util.containers.*
+import com.intellij.util.concurrency.AppExecutorUtil
+import com.intellij.util.containers.filterSmart
+import com.intellij.util.containers.mapSmart
+import com.intellij.util.containers.nullize
+import com.intellij.util.containers.toMutableSmartList
 import com.intellij.util.text.UniqueNameGenerator
+import com.intellij.workspaceModel.ide.WorkspaceModelChangeListener
+import com.intellij.workspaceModel.ide.WorkspaceModelTopics
+import com.intellij.workspaceModel.storage.VersionedStorageChange
+import com.intellij.workspaceModel.storage.bridgeEntities.api.ContentRootEntity
+import com.intellij.workspaceModel.storage.bridgeEntities.api.SourceRootEntity
 import org.jdom.Element
 import org.jetbrains.annotations.TestOnly
+import java.util.concurrent.Callable
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.ConcurrentMap
 import java.util.concurrent.atomic.AtomicBoolean
 import java.util.concurrent.atomic.AtomicReference
 import java.util.concurrent.locks.ReentrantReadWriteLock
 import javax.swing.Icon
-import kotlin.collections.HashMap
 import kotlin.collections.component1
 import kotlin.collections.component2
 import kotlin.concurrent.read
@@ -138,7 +148,7 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
   // When readExternal not all configuration may be loaded, so we need to remember the selected configuration
   // so that when it is eventually loaded, we can mark is as a selected.
   // See also notYetAppliedInitialSelectedConfigurationId, which helps when the initially selected RC is stored in some arbitrary *.run.xml file in project
-  private var selectedConfigurationId: String? = null
+  protected open var selectedConfigurationId: String? = null
   // RCs stored in arbitrary *.run.xml files are loaded a bit later than RCs from workspace and from .idea/runConfigurations.
   // This var helps if initially selected RC is a one from such file.
   // Empty string means that there's no information about initially selected RC in workspace.xml => IDE should select any.
@@ -204,11 +214,13 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
 
   init {
     val messageBusConnection = project.messageBus.connect()
-    messageBusConnection.subscribe(ProjectTopics.PROJECT_ROOTS, object : ModuleRootListener {
-      override fun rootsChanged(event: ModuleRootEvent) {
-        clearSelectedConfigurationIcon()
+    WorkspaceModelTopics.getInstance(project).subscribeAfterModuleLoading(messageBusConnection, object : WorkspaceModelChangeListener {
+      override fun changed(event: VersionedStorageChange) {
+        if (event.getChanges(ContentRootEntity::class.java).isNotEmpty() || event.getChanges(SourceRootEntity::class.java).isNotEmpty()) {
+          clearSelectedConfigurationIcon()
 
-        deleteRunConfigsFromArbitraryFilesNotWithinProjectContent()
+          deleteRunConfigsFromArbitraryFilesNotWithinProjectContent()
+        }
       }
     })
 
@@ -225,6 +237,14 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
     })
 
     BeforeRunTaskProvider.EP_NAME.getPoint(project).addChangeListener(stringIdToBeforeRunProvider::drop, project)
+  }
+
+  override fun shouldSetRunConfigurationFromContext(): Boolean {
+    return Registry.`is`("select.run.configuration.from.context") && !isRunWidgetActive()
+  }
+
+  override fun isRunWidgetActive(): Boolean {
+    return RunToolbarSlotManager.getInstance(project).active
   }
 
   private fun clearSelectedConfigurationIcon() {
@@ -343,9 +363,17 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
   }
 
   private fun deleteRunConfigsFromArbitraryFilesNotWithinProjectContent() {
-    val deletedConfigs = lock.write { rcInArbitraryFileManager.findRunConfigsThatAreNotWithinProjectContent() }
-    // don't delete file just because it has become excluded
-    removeConfigurations(deletedConfigs, deleteFileIfStoredInArbitraryFile = false)
+    ReadAction
+      .nonBlocking(Callable {
+        lock.read { rcInArbitraryFileManager.findRunConfigsThatAreNotWithinProjectContent() }
+      })
+      .coalesceBy(this)
+      .expireWith(project)
+      .finishOnUiThread(ModalityState.defaultModalityState()) {
+        // don't delete file just because it has become excluded
+        removeConfigurations(it, deleteFileIfStoredInArbitraryFile = false)
+      }
+      .submit(AppExecutorUtil.getAppExecutorService())
   }
 
   // Paths in <code>deletedFilePaths</code> and <code>updatedFilePaths</code> may be not related to the project, use ProjectIndex.isInContent() when needed
@@ -357,7 +385,7 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
     removeConfigurations(deletedRunConfigs, deleteFileIfStoredInArbitraryFile = false)
 
     for (filePath in updatedFilePaths) {
-      val deletedAndAddedRunConfigs = lock.write { rcInArbitraryFileManager.loadChangedRunConfigsFromFile(this, filePath) }
+      val deletedAndAddedRunConfigs = lock.read { rcInArbitraryFileManager.loadChangedRunConfigsFromFile(this, filePath) }
 
       for (runConfig in deletedAndAddedRunConfigs.addedRunConfigs) {
         addConfiguration(runConfig)
@@ -407,8 +435,6 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
     val newId = settings.uniqueID
     var existingId: String?
     lock.write {
-      listManager.immutableSortedSettingsList = null
-
       // https://youtrack.jetbrains.com/issue/IDEA-112821
       // we should check by instance, not by id (todo is it still relevant?)
       existingId = if (idToSettings.get(newId) === settings) newId else findExistingConfigurationId(settings)
@@ -419,10 +445,13 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
           if (selectedConfigurationId == it) {
             selectedConfigurationId = newId
           }
+
+          listManager.updateConfigurationId(it, newId)
         }
       }
 
       idToSettings.put(newId, settings)
+      listManager.requestSort()
 
       if (existingId == null) {
         refreshUsagesList(settings)
@@ -574,6 +603,11 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
       eventPublisher.runConfigurationSelected(value)
     }
 
+  internal fun isFileContainsRunConfiguration(file: VirtualFile): Boolean {
+    val runConfigs = lock.read { rcInArbitraryFileManager.getRunConfigsFromFiles(listOf(file.path)) }
+    return runConfigs.isNotEmpty()
+  }
+
   internal fun selectConfigurationStoredInFile(file: VirtualFile) {
     val runConfigs = lock.read { rcInArbitraryFileManager.getRunConfigsFromFiles(listOf(file.path)) }
     runConfigs.find { idToSettings.containsKey(it.uniqueID) }?.let { selectedConfiguration = it }
@@ -679,22 +713,31 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
     return methodElement
   }
 
-  @Suppress("unused")
   /**
    * used by MPS. Do not use if not approved.
    */
   fun reloadSchemes() {
+    var arbitraryFilePaths: Collection<String>
     lock.write {
       // not really required, but hot swap friendly - 1) factory is used a key, 2) developer can change some defaults.
       templateDifferenceHelper.clearCache()
       templateIdToConfiguration.clear()
       listManager.idToSettings.clear()
+      arbitraryFilePaths = rcInArbitraryFileManager.clearAllAndReturnFilePaths()
       recentlyUsedTemporaries.clear()
 
       stringIdToBeforeRunProvider.drop()
     }
+
     workspaceSchemeManager.reload()
     projectSchemeManager.reload()
+    reloadRunConfigsFromArbitraryFiles(arbitraryFilePaths)
+  }
+
+  private fun reloadRunConfigsFromArbitraryFiles(filePaths: Collection<String>) {
+    for (filePath in filePaths) {
+      updateRunConfigsFromArbitraryFiles(emptyList(), filePaths)
+    }
   }
 
   protected open fun addExtensionPointListeners() {
@@ -747,7 +790,7 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
   }
 
   override fun loadState(parentNode: Element) {
-    config.migrateToRegistry()
+    config.migrateToAdvancedSettings()
     val oldSelectedConfigurationId: String?
     val isFirstLoadState = isFirstLoadState.compareAndSet(true, false)
     if (isFirstLoadState) {
@@ -848,8 +891,12 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
     if (selectedConfiguration == null) {
       // Empty string means that there's no information about initially selected RC in workspace.xml => IDE should select any.
       notYetAppliedInitialSelectedConfigurationId = selectedConfigurationId ?: ""
-      selectedConfiguration = allSettings.firstOrNull { it.type.isManaged }
+      selectAnyConfiguration()
     }
+  }
+
+  private fun selectAnyConfiguration() {
+    selectedConfiguration = allSettings.firstOrNull { it.type.isManaged }
   }
 
   fun readContext(parentNode: Element) {
@@ -1019,7 +1066,7 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
 
     tempConfiguration.isTemporary = true
     addConfiguration(tempConfiguration)
-    if (Registry.`is`("select.run.configuration.from.context")) {
+    if (shouldSetRunConfigurationFromContext()) {
       selectedConfiguration = tempConfiguration
     }
   }
@@ -1251,7 +1298,7 @@ open class RunManagerImpl @JvmOverloads constructor(val project: Project, shared
     }
 
     if (selectedConfigurationWasRemoved) {
-      selectedConfiguration = null
+      selectAnyConfiguration()
     }
 
     removed.forEach { eventPublisher.runConfigurationRemoved(it) }
