+++ plugins/InspectionGadgets/src/com/siyeh/ig/fixes/IntroduceHolderFix.java
@@ -1,4 +1,4 @@
-// Copyright 2000-2021 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+// Copyright 2000-2022 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
 package com.siyeh.ig.fixes;
 
 import com.intellij.codeInsight.CodeInsightUtilCore;
@@ -15,13 +15,14 @@ import com.intellij.psi.codeStyle.VariableKind;
 import com.intellij.psi.search.searches.ReferencesSearch;
 import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.psi.util.PsiUtil;
-import com.intellij.refactoring.extractMethod.ExtractMethodProcessor;
 import com.intellij.refactoring.rename.RenamePsiElementProcessor;
 import com.intellij.refactoring.rename.inplace.MemberInplaceRenamer;
+import com.intellij.util.CommonJavaRefactoringUtil;
 import com.siyeh.InspectionGadgetsBundle;
 import com.siyeh.ig.InspectionGadgetsFix;
 import com.siyeh.ig.psiutils.CommentTracker;
 import com.siyeh.ig.psiutils.ControlFlowUtils;
+import com.siyeh.ig.psiutils.VariableAccessUtils;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
 
@@ -134,12 +135,16 @@ public class IntroduceHolderFix extends InspectionGadgetsFix {
     }
 
     final PsiExpression holderReference = elementFactory.createExpressionFromText(holderName + "." + fieldName, field);
-    for (PsiReference reference : ReferencesSearch.search(field).findAll()) {
-      reference.getElement().replace(holderReference);
+    final PsiClass containingClass = PsiUtil.getTopLevelClass(field);
+    if (containingClass == null) return;
+    // Search references within top-level class only.
+    // If there are other references, the fix should not be created, see isStaticAndAssignedOnce
+    for (PsiReferenceExpression reference : VariableAccessUtils.getVariableReferences(field, containingClass)) {
+      reference.replace(holderReference);
     }
     field.delete();
 
-    if (isOnTheFly()) {
+    if (isOnTheFly() && containingClass.isPhysical()) {
       invokeInplaceRename(holderClass, holderName, suggestHolderName(field));
     }
   }
@@ -244,6 +249,6 @@ public class IntroduceHolderFix extends InspectionGadgetsFix {
     final PsiElement[] elements = rhs == null ? PsiElement.EMPTY_ARRAY : new PsiElement[] {rhs};
     final HashSet<PsiField> usedFields = new HashSet<>();
     final PsiClass targetClass = field.getContainingClass();
-    return ExtractMethodProcessor.canBeStatic(targetClass, expressionStatement, elements, usedFields) && usedFields.isEmpty();
+    return CommonJavaRefactoringUtil.canBeStatic(targetClass, expressionStatement, elements, usedFields) && usedFields.isEmpty();
   }
 }
